Fix all examples, resolves #55

+++ Examples/Calibration/recorder_realsense_D435i.cc
@@ -28,8 +28,7 @@
 
 #include <condition_variable>
 
-#include <opencv2/core/core.hpp>
-#include <opencv2/highgui.hpp>
+#include <opencv2/opencv.hpp>
 
 #include <librealsense2/rs.hpp>
 #include "librealsense2/rsutil.h"
@@ -267,11 +266,11 @@ int main(int argc, char **argv) {
         //assert(vAccel.size() == vAccel_times.size());
         //assert(vGyro.size() == vGyro_times.size());
 
-        for(int i=0; i<vAccel.size(); ++i){
+        for(size_t i=0; i<vAccel.size(); ++i){
             accFile << std::setprecision(15) << vAccel_times[i] << "," << vAccel[i].x << "," << vAccel[i].y << "," << vAccel[i].z << endl;
         }
 
-        for(int i=0; i<vGyro.size(); ++i){
+        for(size_t i=0; i<vGyro.size(); ++i){
             gyroFile << std::setprecision(15) << vGyro_times[i] << "," << vGyro[i].x << "," << vGyro[i].y << "," << vGyro[i].z << endl;
         }
